All classes connected with REST API moved to package ru.gwp.api [=] BaseApiSteps.java and BaseApiStepsImpl.java [=] ContentType.java, LogLevel.java, RequestMethod.java and Header.java [=] BaseRestApi.java [=] AllureRestAssuredBeforeAll.java [=] RestRequest.java and RestAssuredRequest.java [=] RestAssuredResponse.java [=] TheCatApiImpl.java, TheCatApiSteps.java, TheCatApiStepsImpl.java and TheCatApiTest.java
