+++ src/main/java/ru/gwp/api/BaseRestApi.java
@@ -1,9 +1,9 @@
 package ru.gwp.api;
 
+import ru.gwp.api.constants.ContentType;
 import ru.gwp.api.rest.RestRequest;
-import ru.gwp.constants.ContentType;
 
-import static ru.gwp.constants.ContentType.JSON;
+import static ru.gwp.api.constants.ContentType.JSON;
 
 /** Provides functionality to work with REST API. */
 public interface BaseRestApi {
