+++ src/main/java/ru/gwp/api/constants/LogLevel.java
@@ -1,4 +1,4 @@
-package ru.gwp.constants;
+package ru.gwp.api.constants;
 
 import lombok.Getter;
 import ru.gwp.error.AutoTestError;
